requirements.txt
python-dotenv
tiktoken
openai
pinecone-client
PyPDF2
tqdm
jsonpickle
regex
requests
beautifulsoup4
nltk
rake-nltk
pytz
python-telegram-bot
pytesseract
pillow
pdf2image
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
wolframalpha
nest_asyncio
